move login_user to one location
